+++ packages/zosfiles/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Zowe z/OS files SDK package will be documented in thi
 
 ## Recent Changes
 
-- Enhancement: Added support for `IDownloadOptions.extensionMap` to the `Download.allMembers` method to allow for specifying the file extension for each downloaded member. [#2507](https://github.com/zowe/zowe-cli/pull/2507)
+- Enhancement: Added support for the `IDownloadOptions.extensionMap` property to the `Download.allMembers` method to allow for specifying the file extension for each downloaded member. [#2507](https://github.com/zowe/zowe-cli/pull/2507)
 
 ## `8.18.0`